pyproject.toml
[tool.poetry]
name = "fish-coins-bot-master"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
nonebot2 = {extras = ["fastapi"], version = "^2.4.0"}
uvicorn = "^0.32.0"
nonebot-adapter-onebot = "^2.4.6"
nonebot-plugin-apscheduler = "^0.5.0"
tortoise-orm = {extras = ["aiomysql"], version = "^0.22.0"}
httpx = "^0.27.2"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"
playwright = "^1.49.0"
jinja2 = "^3.1.4"
bilibili-api-python = "^17.1.4"
minio = "^7.2.16"
yt-dlp = "^2026.1.31"
jieba = "^0.42.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.nonebot]
plugin_dirs = ["fish_coins_bot/plugins"]
